Create Lock conversation on this commit
Locking the conversation means:

Other users can’t add new comments to this commit.
You and other collaborators with access to this repository can still leave comments that others can see.
You can always unlock this commit again in the future.

We appreciate your understanding and apologize if this action was taken in error. Our automated systems help us manage the large volume of issues we receive, but we know they're not perfect.